Deduplicate TR3 logic
